requirements.txt
selenium==3.141.0
python-decouple==3.1